Initial implementation of eval. Change internal representation of cons to closure.
In R, changing the value of a supplied argument within a function will not affect the value of the variable in the calling frame. We can make an immutable object with a closure and <<- operator. Using closures has another advantage. R doesn't support address comparison of objects but R can compare closures. The expression identical(list(1), list(1)) returns TRUE but identical(<closure1>, <closure2>) returns FALSE.

R has a form of lazy evaluation of function arguments. Arguments are not evaluated until needed. We need to be careful because this may make an infinite list implicitly. See makeCons.